refactor: split user controller by responsibility (#15)
