+++ server/controllers/follow.controller.js
@@ -1,0 +1,63 @@
+import User from '../models/user.model'
+import errorHandler from './../helpers/dbErrorHandler'
+
+const addFollowing = async (req, res, next) => {
+  try{
+    await User.findByIdAndUpdate(req.body.userId, {$push: {following: req.body.followId}})
+    next()
+  }catch(err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
+}
+
+const addFollower = async (req, res) => {
+  try{
+    let result = await User.findByIdAndUpdate(req.body.followId, {$push: {followers: req.body.userId}}, {new: true})
+                            .populate('following', '_id name')
+                            .populate('followers', '_id name')
+                            .exec()
+      result.hashed_password = undefined
+      result.salt = undefined
+      res.json(result)
+    }catch(err) {
+      return res.status(400).json({
+        error: errorHandler.getErrorMessage(err)
+      })
+    }
+}
+
+const removeFollowing = async (req, res, next) => {
+  try{
+    await User.findByIdAndUpdate(req.body.userId, {$pull: {following: req.body.unfollowId}})
+    next()
+  }catch(err) {
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
+}
+
+const removeFollower = async (req, res) => {
+  try{
+    let result = await User.findByIdAndUpdate(req.body.unfollowId, {$pull: {followers: req.body.userId}}, {new: true})
+                            .populate('following', '_id name')
+                            .populate('followers', '_id name')
+                            .exec()
+    result.hashed_password = undefined
+    result.salt = undefined
+    res.json(result)
+  }catch(err){
+      return res.status(400).json({
+        error: errorHandler.getErrorMessage(err)
+      })
+  }
+}
+
+export default {
+  addFollowing,
+  addFollower,
+  removeFollowing,
+  removeFollower
+}
