+++ server/controllers/people.controller.js
@@ -1,0 +1,19 @@
+import User from '../models/user.model'
+import errorHandler from './../helpers/dbErrorHandler'
+
+const findPeople = async (req, res) => {
+  let following = req.profile.following
+  following.push(req.profile._id)
+  try {
+    let users = await User.find({ _id: { $nin : following } }).select('name')
+    res.json(users)
+  }catch(err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
+}
+
+export default {
+  findPeople
+}
